Refactor library to use memory mapping (#1)
This change reworks the original library that loaded the entire library into memory to use memory-mapping. This allows the library to work within devices and operating systems that limit the amount of memory in use i.e. iOS.

It also removes the specific classes that are used for GeoLite as these are implementation specific and for the consumer to create. As a result of this the existent MMDB file now takes a URL for a file for it to be read - ensuring the .alwaysMapped option on the data object.

Due to timing constraints with required projects - the bad test cases are currently commented out. These will be corrected and remediated in due time.